WIP: Remove the notion of QtApp.
